move mixin accessor interfaces to same package
