+++ src/main/java/gregtech/mixin/interfaces/IBlockStemAccesor.java
@@ -1,4 +1,4 @@
-package kubatech.api;
+package gregtech.mixin.interfaces;
 
 import net.minecraft.block.Block;
